% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpData_dirs.R
\name{harpData_info}
\alias{harpData_info}
\title{Get info about data in the harpData package}
\usage{
harpData_info(
  file_format = c("vfld", "grib", "netcdf", "fctable", "vobs", "obstable"),
  fctable_fcst_type = c("det", "ens")
)
}
\arguments{
\item{file_format}{The file format in harpData that you want information for.}

\item{fctable_fcst_type}{In the case of \code{file_format = "fctable"}, the
data are separated into deterministic and ensemble groups. Set to "det" to
get the info for deterministic data and "ens" for ensemble data.}
}
\value{
A named list giving information about directories, start and end
  dates, available forecast models for the file format and in the case of
  ensemble data, the members and where appropriate the lags of the members.
}
\description{
Get info about data in the harpData package
}
\examples{
harpData_dirs("vfld")
harpData_dirs("grib")
harpData_dirs("netcdf")
harpData_dirs("fctable", "det")
harpData_dirs("fctable", "ens")
harpData_dirs("vobs")
harpData_dirs("obstable")

}
